Import dat z data stream (#220)
* Add service which downloads data in .csv format

Add service which downloads zipped .csv used by Internet Stream.
Change datatype of event from Int to Long.

* Use camel case

Ensure that camelCase convention is in the service.

* Refactor InternetStreamService

Remove try, catch construct. Switch to kotlin approach with use function.
Remove for loops. Switch to Kotlin approach with map function.
Make configuration of InternetStreamService more robust.

* Remove redundant buffer.

* Add test for service which downloads data in .csv format

* Simplify InternetStreamService

Simplify test. Add configurable delay after which is the service
started.

+++ server-kotlin/src/main/kotlin/digital/cesko/city_sync/model/CityExport.kt
@@ -98,7 +98,7 @@ data class CityExport(
         val paragraph: Int?,
         val item: Int?,
         val unit: Int?,
-        val event: Int?,
+        val event: Long?,
         val amount: BigDecimal?,
         @JsonDeserialize(using = LocalDateDeserializer::class)
         val date: LocalDate?,
